+++ static/AAPL_dataset.csv
@@ -6369,3 +6369,15 @@ date,Open,High,Low,Close,Volume,EMA20,EMA50,EMA100,EMA200
 2025-02-25,248.0,250.0,244.91,247.04,48013272.0,240.039861446502,238.149090531999,234.42195282978793,224.2800140968045
 2025-02-26,244.33,244.98,239.13,240.36,44433564.0,240.07035083254942,238.2357928640775,234.53953792226739,224.4400139565378
 2025-02-27,239.41,242.46,237.06,237.3,41153639.0,239.80650789611613,238.1990951047019,234.59420053766803,224.56797401667174
+2025-02-28,236.95,242.09,230.2,241.84,56833360.0,240.00017381077174,238.34187568883124,234.73768171513993,224.73983497172975
+2025-03-03,241.79,244.0272,236.112,238.03,47183985.0,239.81253820974587,238.3296452696614,234.8028761366223,224.87207541977224
+2025-03-04,237.705,240.07,234.68,235.93,53798062.0,239.44277266596055,238.23554153359626,234.82519542104563,224.98210452007302
+2025-03-05,235.42,236.55,229.23,235.74,47227643.0,239.09012765015478,238.13767715972975,234.84331036320313,225.08914825619172
+2025-03-06,234.435,237.86,233.1581,235.33,45170419.0,238.73202025490193,238.02757217307368,234.85294778175356,225.19104727851817
+2025-03-07,235.105,241.37,234.76,239.07,46273565.0,238.76420880205413,238.06845169569826,234.93645376627327,225.32914631057272
+2025-03-10,235.54,236.16,224.22,227.48,71451281.0,237.68952224947753,237.65321829586694,234.78880121644607,225.3505478398208
+2025-03-11,223.805,225.8399,217.45,220.84,76137410.0,236.08480584476538,236.99387640191136,234.51258733097188,225.30566676678774
+2025-03-12,220.14,221.75,214.91,216.98,62547467.0,234.26530052621632,236.20901850379718,234.16540738382395,225.22282431139683
+2025-03-13,215.95,216.8394,208.42,209.68,61368330.0,231.92384333324335,235.1686648369816,233.680547831669,225.06816934312423
+2025-03-14,211.25,213.95,209.58,213.49,60107582.0,230.1682392062678,234.3185211178843,233.28073500331914,224.9529636780185
+2025-03-17,213.31,215.22,209.97,214.0,48073426.0,228.62840690090897,233.52171636816337,232.89893827058015,224.8439789648044
