+++ static/GOOGL_dataset.csv
@@ -5175,3 +5175,4 @@ date,Open,High,Low,Close,Volume,EMA20,EMA50,EMA100,EMA200
 2025-03-12,166.58,167.6399,163.53,167.11,28372396.0,174.92801906267667,180.99243126214435,180.1476266881324,176.74709128394448
 2025-03-13,166.035,166.13,162.11,162.76,31756214.0,173.76916010432652,180.27743395774652,179.80331724876342,176.6079162462933
 2025-03-14,163.27,166.49,162.45,165.49,31995894.0,172.98066866581922,179.69753458685452,179.51988522403545,176.4972902139919
+2025-03-17,165.03,166.3,163.67,164.29,31184335.0,172.1529859357412,179.09331754423278,179.21830333841098,176.37582463972333
