+++ templates/register.html
@@ -1,0 +1,125 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Register - Stock Trend Prediction</title>
+    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/css/bootstrap.min.css" />
+    <style>
+        body {
+            height: 100vh;
+            margin: 0;
+            display: flex;
+            flex-direction: column;
+            background: #0d1117;
+            color: #58a6ff;
+            font-family: "Orbitron", sans-serif;
+            overflow: hidden;
+            position: relative;
+        }
+
+        .navbar {
+            position: fixed;
+            top: 0;
+            left: 0;
+            width: 100%;
+            display: flex;
+            justify-content: space-between;
+            padding: 15px 30px;
+            background: rgba(13, 17, 23, 0.9);
+            backdrop-filter: blur(10px);
+            z-index: 1000;
+        }
+
+        .navbar a {
+            color: #58a6ff;
+            text-decoration: none;
+            font-weight: bold;
+            transition: 0.3s ease-in-out;
+        }
+
+        .navbar a:hover {
+            color: #79c0ff;
+            text-shadow: 0 0 10px #79c0ff;
+        }
+
+        .content-wrapper {
+            display: flex;
+            flex: 1;
+            margin-top: 60px;
+        }
+
+        .right-content {
+            flex: 1;
+            display: flex;
+            align-items: center;
+            justify-content: center;
+            padding: 20px;
+        }
+
+        .register-form {
+            width: 100%;
+            max-width: 350px;
+        }
+
+        .form-control {
+            background: #010409;
+            border: 1px solid #58a6ff;
+            color: #58a6ff;
+            box-shadow: 0 0 10px #58a6ff inset;
+        }
+
+        .btn-primary {
+            background: #58a6ff;
+            border: none;
+            color: #0d1117;
+            font-weight: bold;
+            text-transform: uppercase;
+            box-shadow: 0 0 20px #58a6ff;
+            transition: 0.3s;
+        }
+
+        .btn-primary:hover {
+            background: #79c0ff;
+            box-shadow: 0 0 35px #79c0ff;
+        }
+    </style>
+</head>
+
+<body>
+    <nav class="navbar">
+        <a href="/">Home</a>
+        <a href="/login">Login</a>
+    </nav>
+
+    <div class="content-wrapper">
+        <div class="right-content">
+            <form class="register-form" method="POST">
+                <h2 class="text-center mb-4">Register</h2>
+                {% if error %}
+                <div class="alert alert-danger">{{ error }}</div>
+                {% endif %}
+                <div class="mb-3">
+                    <label for="username" class="form-label">Username</label>
+                    <input type="text" id="username" name="username" class="form-control" required />
+                </div>
+                <div class="mb-3">
+                    <label for="password" class="form-label">Password</label>
+                    <input type="password" id="password" name="password" class="form-control" required />
+                </div>
+                <div class="mb-3">
+                    <label for="confirm_password" class="form-label">Confirm Password</label>
+                    <input type="password" id="confirm_password" name="confirm_password" class="form-control"
+                        required />
+                </div>
+                <button type="submit" class="btn btn-primary w-100">Register</button>
+                <div class="text-center mt-3">
+                    <a href="/login" class="text-info">Already have an account? Login</a>
+                </div>
+            </form>
+        </div>
+    </div>
+</body>
+
+</html>
