+++ templates/tools.html
@@ -134,19 +134,51 @@
             background: #00adb5;
             box-shadow: 0 0 25px #00adb5;
         }
+
+        .user-icon {
+            text-decoration: none;
+            display: flex;
+            align-items: center;
+        }
+
+        .user-circle {
+            width: 40px;
+            height: 40px;
+            background: #58a6ff;
+            color: #0d1117;
+            font-weight: bold;
+            font-size: 20px;
+            border-radius: 50%;
+            display: flex;
+            justify-content: center;
+            align-items: center;
+            box-shadow: 0 0 10px #58a6ff;
+            transition: box-shadow 0.3s;
+        }
+
+        .user-circle:hover {
+            box-shadow: 0 0 20px #79c0ff;
+        }
     </style>
 </head>
 
 <body>
     <nav class="navbar">
         <a href="/">Home</a>
-        <a href="#">About</a>
-        <a href="/tool">Use the Tool</a>
+        {% if session['user'] %}
+        <a href="{{ url_for('auth_app.profile') }}" class="user-icon">
+            <div class="user-circle">{{ session['user'][0]|upper }}</div>
+        </a>
+        {% else %}
+        <a href="{{ url_for('auth_app.login') }}" class="user-icon">
+            <div class="user-circle">L</div>
+        </a>
+        {% endif %}
     </nav>
 
     <div class="container">
         <h1>Stock Trend Prediction</h1>
-        <form method="POST">
+        <form method="POST" enctype="multipart/form-data">
             <div class="row">
                 <div class="col-md-8 mb-3">
                     <label for="stock" class="form-label">Stock Ticker:</label>
@@ -204,13 +236,28 @@
 
         {% if dataset_link %}
         <div class="mt-4">
-            <a href="{{ url_for('download_file', filename=dataset_link.split('/')[-1]) }}" class="btn btn-success"
-                download>Download Dataset (CSV)</a>
+            <a href="{{ url_for('tool_app.download_file', filename=dataset_link.split('/')[-1]) }}"
+                class="btn btn-success" download>
+                Download Dataset (CSV)
+            </a>
         </div>
         {% endif %}
     </div>
 
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha1/dist/js/bootstrap.bundle.min.js"></script>
+
+    <script>
+        document.querySelector("form").addEventListener("submit", function (event) {
+            const startDate = document.getElementById("start_date").value;
+            const endDate = document.getElementById("end_date").value;
+
+            if (!startDate || !endDate) {
+                event.preventDefault();
+                alert("Both start date and end date are required.");
+            }
+        });
+    </script>
+
 </body>
 
 </html>
